requirements.txt

Flask==1.1.2

numpy==1.18.5

Pillow==7.1.2



tensorflow==2.3.1


